refactor: 영상통화방 컴포넌트 분리
